requirements.txt
streamlit
python-dotenv
PyPDF2
langchain
langchain-community
langchain-huggingface
transformers
torch
sentence-transformers
chromadb
